Just make it a test
